Fix partitions bigger than 2TB
Partitions which are bigger than 2TB doesn't fit unsigned long,
change it ot unsigned long long.

+++ yaboot/include/partition.h
@@ -36,8 +36,8 @@ enum disk_label {
 struct partition_t {
 	struct partition_t*	next;
 	int			part_number;
-	unsigned long  		part_start; /* In blocks */
-	unsigned long  		part_size; /* In blocks */
+	unsigned long long	part_start; /* In blocks */
+	unsigned long long	part_size; /* In blocks */
 	unsigned short		blocksize;
 	enum disk_label		label;
 	int			sys_ind; /* fs type */
